backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
python-dotenv==1.0.0
openai==1.40.0
supabase==2.5.0
apscheduler==3.10.4
httpx==0.27.0
